# openapi.yaml
openapi: 3.1.0
info:
  title: Food Bank API
  version: '1.0.0'
  description: |
    JSON API for running a food bank: referrals in, sessions scheduled, pick
    lists produced, stock adjusted.

    **Read `API.md` alongside this.** A schema describes shapes; it cannot
    describe the login/refresh cycle, what happens to a referral from an
    unrecognised address, or why a team lead receives fewer fields than an
    admin. Those are the parts most likely to be got wrong.

    ## Generating a client

    ```
    npx openapi-typescript openapi.yaml -o src/api/schema.d.ts
    ```
servers:
  - url: http://127.0.0.1:8787
    description: Local wrangler dev
  - url: https://foodbank-server-production.workers.dev
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/sessions/{sessionId}/sms-reminders:
    parameters: [{ $ref: '#/components/parameters/SessionId' }]
    post:
      tags: [SMS]
      summary: Text every household on this session that has not been texted
      description: |
        The **Send SMS Reminders** button on the run-session screen. Team leader
        or admin. No body.

        **Who gets one:** every household holding a place — `pending_review`,
        `active` and `reviewed` alike. Pick-list generation uses the same rule,
        so the run-session client has a named parcel for every household it can
        remind. Cancelled and rejected households are left out.

        **Sending twice is safe.** A household already reminded for this session
        is skipped however many times the button is pressed, so a second press
        reaches only households referred since the first — and households whose
        first attempt failed.

        **A failure does not count as reminded.** No number held, a number that
        turns out not to be a mobile, or the provider refusing it all produce a
        `failure` message against that household and leave them unreminded, so
        the next press tries again. A permanently wrong number therefore fails
        on every press; that is the food bank being told the number is no good.

        Two wordings, and the client chooses neither: a collection reminder
        carries the date, the time and the place, a delivery reminder carries
        the date and the session's `deliveryTime` and **no address**.
      responses:
        '200':
          description: What happened
          content:
            application/json:
              schema:
                type: object
                required: [reminded, failed, alreadyReminded]
                properties:
                  reminded: { type: integer, description: Texts that went. }
                  failed:
                    type: integer
                    description: >-
                      Households left unreminded, each with a `failure` message
                      on their line saying why.
                  alreadyReminded:
                    type: integer
                    description: Skipped because they had already been texted for this session.
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/sessions/{sessionId}/sms-summary:
    parameters: [{ $ref: '#/components/parameters/SessionId' }]
    get:
      tags: [SMS]
      summary: Message counts for the run-session screen
      description: |
        **This is the polling target.** Cheap enough to call every few seconds
        while the screen is open, which is how the unread total and the
        per-household counts stay current. There is no push channel.

        `unreadTotal` is the number the screen is meant to make prominent.

        **Counts are of what came in** — replies and failures — and never of
        what the food bank sent. A count that went up when staff answered a
        message would be counting their own work back at them.
      responses:
        '200':
          description: Counts
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, unreadTotal, households]
                properties:
                  sessionId: { type: string, format: uuid }
                  unreadTotal: { type: integer }
                  households:
                    type: array
                    items:
                      type: object
                      required: [referralId, reminderSentAt, messageCount, unreadCount]
                      properties:
                        referralId: { type: string, format: uuid }
                        reminderSentAt:
                          type: [string, 'null']
                          format: date-time
                          description: Null means this household has not been texted.
                        messageCount: { type: integer }
                        unreadCount: { type: integer }
  /api/v1/referrals/{id}/sms-messages:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [SMS]
      summary: One household's conversation
      description: >
        The whole thread, oldest first, **both directions** — the reminder that
        went, the household's replies, any failures, and anything staff sent
        back. Read-only: it marks nothing read, so it is safe to prefetch and
        safe to retry.

        Open to a team leader, which is a deliberate exception of the same kind
        as the listener sheet: a household may text something as personal as
        their reason for referral, and the person running the session is the one
        who can act on it.
      responses:
        '200':
          description: The thread
          content:
            application/json:
              schema:
                type: object
                required: [referralId, messages]
                properties:
                  referralId: { type: string, format: uuid }
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/SmsMessage' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [SMS]
      summary: Text this household back
      description: >
        Sends a reply and records it on the thread. It does not change any
        count — the counts are of what came in.

        **The no-identifying-data rule applies to what a human types here.** The
        provider is given the number and nothing that says whose it is, so a
        reply must not carry the household's name or address. No code can
        enforce that on free text; say so on the screen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 480 }
      responses:
        '200':
          description: Sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SmsMessage' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: >-
            The household has no number, sending is not configured, or the
            provider would not take it. **Nothing is recorded** — unlike a
            reminder, a staff reply that did not go leaves no `failure` on the
            thread, because it would read as something the household should see.
  /api/v1/referrals/{id}/sms-messages/read:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [SMS]
      summary: Mark this household's messages read
      description: >
        Fire this when the user expands the household's line — opening the
        messages is what marks them read, and there is nothing further for
        anybody to press.

        Separate from the `GET` on purpose: a `GET` that writes gets retried by
        browsers and cleared by any prefetch, and somebody's unread count would
        vanish without anyone having looked.

        Idempotent. Only inbound messages are ever unread, so this never touches
        what the food bank sent.
      responses:
        '200':
          description: Marked
          content:
            application/json:
              schema:
                type: object
                required: [markedRead]
                properties:
                  markedRead: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/sms-messages/{id}/read:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [SMS]
      summary: Mark one loose reply read
      description: "Admin only. The unmatched screen's equivalent of expanding a household."
      responses:
        '200':
          description: Marked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SmsMessage' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <accessToken>`. Expires after 15 minutes, or
        when the eight-hour sign-in ends if that comes first; see the refresh
        flow in API.md.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    SessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: >
        Every failure has this shape. `requestId` also appears in the
        `x-request-id` header — quote it when reporting a problem.
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                [
                  BAD_REQUEST,
                  VALIDATION_FAILED,
                  UNAUTHORIZED,
                  FORBIDDEN,
                  NOT_FOUND,
                  CONFLICT,
                  UNPROCESSABLE,
                  INTERNAL_ERROR,
                ]
            message:
              type: string
              description: Safe to show a user. Never contains personal data.
            details:
              type: object
              description: Structured context, such as which fields failed.
            requestId: { type: string }
...
    SmsMessage:
      type: object
      description: |
        One message on a household's thread, or one loose reply.

        **`body` is personal data of the most sensitive kind here** — a
        household's own words about why they need help. It is deleted thirty
        days after `occurredAt`, along with everything else on this table, so a
        client must not treat a thread as a permanent record.
      required: [id, referralId, kind, body, occurredAt, readAt]
      properties:
        id: { type: string, format: uuid }
        referralId:
          type: [string, 'null']
          format: uuid
          description: Null on a loose reply — a text with no upcoming referral behind it.
        kind:
          type: string
          enum: [reminder, staff_reply, household_reply, failure]
          description: >
            `reminder` — what the food bank sent about the session.

            `household_reply` — what the household texted back. **The only kind
            that is ever unread.**

            `staff_reply` — a person answering from the session screen.

            `failure` — the reminder did not go: no number, a number that is not
            a mobile, or the provider refused it. Not a message anybody sent,
            but it belongs where somebody will see it, and it arrives already
            read because it is not somebody waiting for an answer.
        body:
          type: string
          description: 'For a `failure`, the reason rather than a message.'
        occurredAt: { type: string, format: date-time }
        readAt: { type: [string, 'null'], format: date-time }
        phone:
          type: string
          description: >-
            The household's number, in E.164 where it could be normalised.
            Present on every message: on a loose reply it is the only way to
            act on one, and on a thread it is the same number the referral
            already carries, so withholding it there would buy nothing.
...